+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Fix unset option
 
 ## 1.18.0 - 2026-07-08
 - Sort entity settings table by zone then Eaton name (U1)